Kavitha | refactored provider function constants
