+++ src/features/DisplayControls/NursingTasks/utils/constants.js
@@ -5,9 +5,3 @@ export const items = [
   { id: "stopped", text: "Stopped" },
   { id: "skipped", text: "Skipped" },
 ];
-
-export const performerFunction = "Performer";
-
-export const requesterFunction = "Witness";
-
-export const verifierFunction = "Verifier";
